@@ -1,9 +1,9 @@
 {
     "require-dev": {
-        "phpunit/phpunit": "9.5.11",
-        "psalm/plugin-phpunit": "0.16.1",
-        "symfony/phpunit-bridge": "6.0.0",
-        "vimeo/psalm": "4.17.0"
+        "phpunit/phpunit": "9.5.21",
+        "psalm/plugin-phpunit": "0.17.0",
+        "symfony/phpunit-bridge": "6.1.3",
+        "vimeo/psalm": "4.25.0"
     },
     "config": {
         "allow-plugins": {
