@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Serendipity HQ Value Objects Component.
  *
@@ -13,33 +15,22 @@ namespace SerendipityHQ\Component\ValueObjects\Email;
 
 use Egulias\EmailValidator\EmailValidator;
 use Egulias\EmailValidator\Validation\RFCValidation;
-use Safe\Exceptions\StringsException;
-use function Safe\sprintf;
 use SerendipityHQ\Component\ValueObjects\Common\DisableWritingMethodsTrait;
 
-/**
- * {@inheritDoc}
- */
+use function Safe\sprintf;
+
 final class Email implements EmailInterface
 {
     use DisableWritingMethodsTrait;
 
-    /** @var string */
-    private $email;
-
-    /** @var string */
-    private $mailBox;
+    private string $email;
+    private string $mailBox;
 
     /** @var string */
     private $host;
 
     /**
-     * {@inheritDoc}
-     *
      * @param string $value The email to set in the object
-     *
-     * @throws \InvalidArgumentException
-     * @throws StringsException
      */
     public function __construct($value)
     {
@@ -54,33 +45,26 @@ final class Email implements EmailInterface
         [$this->mailBox, $this->host] = \explode('@', $this->email);
     }
 
-    /**
-     * {@inheritDoc}
-     */
+    public function __toString(): string
+    {
+        return sprintf('%s@%s', $this->mailBox, $this->host);
+    }
+
     public function getEmail(): string
     {
         return $this->email;
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function getHost(): string
     {
         return $this->host;
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function getMailBox(): string
     {
         return $this->mailBox;
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function changeMailBox(string $newMailbox): EmailInterface
     {
         $copy          = clone $this;
@@ -90,21 +74,8 @@ final class Email implements EmailInterface
         return $copy;
     }
 
-    /**
-     * {@inheritDoc}
-     *
-     * @throws StringsException
-     */
     public function toString(array $options = []): string
     {
         return $this->__toString();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function __toString(): string
-    {
-        return sprintf('%s@%s', $this->mailBox, $this->host);
     }
 }
