@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Serendipity HQ Value Objects Component.
  *
@@ -12,11 +14,10 @@
 namespace SerendipityHQ\Component\ValueObjects\Uri;
 
 use Laminas\Uri\Exception\InvalidArgumentException;
-use Laminas\Uri\Exception\InvalidUriPartException;
 use Laminas\Uri\Uri as BaseUri;
-use Safe\Exceptions\StringsException;
-use function Safe\sprintf;
 use SerendipityHQ\Component\ValueObjects\Common\DisableWritingMethodsTrait;
+
+use function Safe\sprintf;
 
 /**
  * Default implementation of an Uri value obeject.
@@ -25,18 +26,10 @@ final class Uri implements UriInterface
 {
     use DisableWritingMethodsTrait;
 
-    /** @var BaseUri */
-    private $valueObject;
+    private BaseUri $valueObject;
 
     /**
-     * {@inheritDoc}
-     *
      * @param string|UriInterface $uri
-     *
-     * @throws StringsException
-     * @throws InvalidArgumentException
-     * @throws InvalidUriPartException
-     * @throws InvalidArgumentException
      */
     public function __construct($uri)
     {
@@ -56,53 +49,41 @@ final class Uri implements UriInterface
             $this->setQuery($uri->getQuery());
             $this->setFragment($uri->getFragment());
         } elseif (null !== $uri) {
-            throw new InvalidArgumentException(sprintf('Expecting a string or a URI object, received "%s"', (\is_object($uri) ? \get_class($uri) : \gettype($uri))));
+            throw new InvalidArgumentException(sprintf('Expecting a string or a URI object, received "%s"', \is_object($uri) ? \get_class($uri) : \gettype($uri)));
         }
     }
 
-    /**
-     * {@inheritDoc}
-     */
+    public function __toString(): string
+    {
+        return $this->valueObject->toString();
+    }
+
     public function getFragment(): ?string
     {
         return $this->valueObject->getFragment();
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function getHost(): ?string
     {
         return $this->valueObject->getHost();
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function getPath(): ?string
     {
         return $this->valueObject->getPath();
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function getPort(): ?int
     {
         return $this->valueObject->getPort();
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function getQuery(): ?string
     {
         return $this->valueObject->getQuery();
     }
 
     /**
-     * {@inheritDoc}
-     *
      * @psalm-suppress MixedReturnTypeCoercion
      *
      * @return array<string,string>
@@ -112,49 +93,31 @@ final class Uri implements UriInterface
         return $this->valueObject->getQueryAsArray();
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function getScheme(): ?string
     {
         return $this->valueObject->getScheme();
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function getUserInfo(): ?string
     {
         return $this->valueObject->getUserInfo();
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function isAbsolute(): bool
     {
         return $this->valueObject->isAbsolute();
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function isValid(): bool
     {
         return $this->valueObject->isValid();
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function isValidRelative(): bool
     {
         return $this->valueObject->isValidRelative();
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function makeRelative($baseUri): UriInterface
     {
         if ($baseUri instanceof UriInterface) {
@@ -166,9 +129,6 @@ final class Uri implements UriInterface
         return $this;
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function normalize(): UriInterface
     {
         $this->valueObject->normalize();
@@ -176,9 +136,6 @@ final class Uri implements UriInterface
         return $this;
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function parse(string $uri): UriInterface
     {
         $this->valueObject->parse($uri);
@@ -186,9 +143,6 @@ final class Uri implements UriInterface
         return $this;
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function setFragment(?string $fragment): UriInterface
     {
         $this->valueObject->setFragment($fragment);
@@ -196,9 +150,6 @@ final class Uri implements UriInterface
         return $this;
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function setHost(?string $host): UriInterface
     {
         $this->valueObject->setHost($host);
@@ -206,9 +157,6 @@ final class Uri implements UriInterface
         return $this;
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function setPath(?string $path): UriInterface
     {
         $this->valueObject->setPath($path);
@@ -216,9 +164,6 @@ final class Uri implements UriInterface
         return $this;
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function setPort(?int $port): UriInterface
     {
         $this->valueObject->setPort($port);
@@ -227,8 +172,6 @@ final class Uri implements UriInterface
     }
 
     /**
-     * {@inheritDoc}
-     *
      * @param array<string,string>|string|null $query
      */
     public function setQuery($query): UriInterface
@@ -238,9 +181,6 @@ final class Uri implements UriInterface
         return $this;
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function setScheme(?string $scheme): UriInterface
     {
         $this->valueObject->setScheme($scheme);
@@ -248,9 +188,6 @@ final class Uri implements UriInterface
         return $this;
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function setUserInfo(?string $userInfo): UriInterface
     {
         $this->valueObject->setUserInfo($userInfo);
@@ -258,9 +195,6 @@ final class Uri implements UriInterface
         return $this;
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function toString(array $options = []): string
     {
         return $this->__toString();
@@ -269,13 +203,5 @@ final class Uri implements UriInterface
     public function getValueObject(): BaseUri
     {
         return $this->valueObject;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function __toString(): string
-    {
-        return $this->valueObject->toString();
     }
 }
