@@ -1,13 +1,25 @@
 <?php
+
 declare(strict_types=1);
+
+/*
+ * This file is part of the Serendipity HQ Value Objects Component.
+ *
+ * Copyright (c) Adamo Aerendir Crespi <aerendir@serendipityhq.com>.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 /**
  * This configuration will be read and overlaid on top of the
  * default configuration. Command line arguments will be applied
  * after this file is read.
  */
+
 return [
     'target_php_version' => '7.4',
-    'minimum_severity' => \Phan\Issue::SEVERITY_LOW,
+    'minimum_severity'   => \Phan\Issue::SEVERITY_LOW,
 
     // A list of directories that should be parsed for class and
     // method information. After excluding the directories
@@ -20,7 +32,7 @@ return [
         'src',
         'tests',
         'vendor',
-        'vendor-bin/phpunit/vendor'
+        'vendor-bin/phpunit/vendor',
     ],
 
     // A directory list that defines files that will be excluded
@@ -39,22 +51,22 @@ return [
         'vendor-bin/phpunit/vendor',
     ],
 
-    'quick_mode' => false,
-    'analyze_signature_compatibility' => true,
-    'allow_missing_properties' => false,
-    'null_casts_as_any_type' => false,
-    'null_casts_as_array' => false,
-    'array_casts_as_null' => false,
-    'scalar_implicit_cast' => false,
+    'quick_mode'                                  => false,
+    'analyze_signature_compatibility'             => true,
+    'allow_missing_properties'                    => false,
+    'null_casts_as_any_type'                      => false,
+    'null_casts_as_array'                         => false,
+    'array_casts_as_null'                         => false,
+    'scalar_implicit_cast'                        => false,
     'ignore_undeclared_variables_in_global_scope' => false,
-    'suppress_issue_types' => [
+    'suppress_issue_types'                        => [
         'PhanUnusedPublicMethodParameter',
         'PhanUnreferencedClass',
         'PhanUnreferencedPublicClassConstant',
         'PhanUnreferencedPublicMethod',
         'PhanUnusedPublicNoOverrideMethodParameter',
         'PhanUnreferencedProtectedMethod',
-        'PhanUnusedPublicFinalMethodParameter'
+        'PhanUnusedPublicFinalMethodParameter',
     ],
 
     // A regular expression to match files to be excluded
@@ -65,9 +77,9 @@ return [
     // can't be removed for whatever reason.
     // (e.g. '@Test\.php$@', or '@vendor/.*/(tests|Tests)/@')
     'exclude_file_regex' => '@^vendor/.*/(stubs?|Stubs?)/@',
-    'plugins' => [
+    'plugins'            => [
         'vendor-bin/phan/vendor/drenso/phan-extensions/Plugin/Annotation/SymfonyAnnotationPlugin.php',
         'vendor-bin/phan/vendor/drenso/phan-extensions/Plugin/DocComment/InlineVarPlugin.php',
-        'vendor-bin/phan/vendor/drenso/phan-extensions/Plugin/DocComment/MethodPlugin.php'
-    ]
+        'vendor-bin/phan/vendor/drenso/phan-extensions/Plugin/DocComment/MethodPlugin.php',
+    ],
 ];
