@@ -1,9 +1,9 @@
 <?php
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 /*
- * This file is part of the Serendipity HQ Aws Ses Bundle.
+ * This file is part of the Serendipity HQ Value Objects Component.
  *
  * Copyright (c) Adamo Aerendir Crespi <aerendir@serendipityhq.com>.
  *
@@ -11,74 +11,17 @@ declare(strict_types = 1);
  * file that was distributed with this source code.
  */
 
-use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
-use Rector\Core\Configuration\Option;
+use Rector\Config\RectorConfig;
 use Rector\Core\ValueObject\PhpVersion;
-use Rector\PHPUnit\Set\PHPUnitSetList;
-use Rector\Set\ValueObject\SetList;
-use Rector\Symfony\Set\SymfonySetList;
+use SerendipityHQ\Integration\Rector\SerendipityHQ;
 
-return static function (ContainerConfigurator $containerConfigurator) : void {
-    $parameters = $containerConfigurator->parameters();
+return static function (RectorConfig $rectorConfig): void {
+    $rectorConfig->phpVersion(PhpVersion::PHP_74);
+    $rectorConfig->paths([__DIR__ . '/src', __DIR__ . '/tests']);
+    $rectorConfig->bootstrapFiles([__DIR__ . '/vendor-bin/phpunit/vendor/autoload.php']);
+    $rectorConfig->import(SerendipityHQ::SHQ_LIBRARY);
 
-    $parameters->set(Option::PHP_VERSION_FEATURES, PhpVersion::PHP_74);
-
-    $parameters->set(Option::PATHS, [
-        __DIR__ . '/src',
-        __DIR__ . '/tests'
-    ]);
-
-    $parameters->set(Option::BOOTSTRAP_FILES, [__DIR__ . '/vendor-bin/phpunit/vendor/autoload.php']);
-
-    $containerConfigurator->import(SetList::CODE_QUALITY);
-    $containerConfigurator->import(SetList::CODING_STYLE);
-    $containerConfigurator->import(SetList::PHP_52);
-    $containerConfigurator->import(SetList::PHP_53);
-    $containerConfigurator->import(SetList::PHP_54);
-    $containerConfigurator->import(SetList::PHP_56);
-    $containerConfigurator->import(SetList::PHP_70);
-    $containerConfigurator->import(SetList::PHP_71);
-    $containerConfigurator->import(SetList::PHP_72);
-    $containerConfigurator->import(SetList::PHP_73);
-    $containerConfigurator->import(PHPUnitSetList::PHPUNIT_40);
-    $containerConfigurator->import(PHPUnitSetList::PHPUNIT_50);
-    $containerConfigurator->import(PHPUnitSetList::PHPUNIT_60);
-    $containerConfigurator->import(PHPUnitSetList::PHPUNIT_70);
-    $containerConfigurator->import(PHPUnitSetList::PHPUNIT_75);
-    $containerConfigurator->import(PHPUnitSetList::PHPUNIT_80);
-    $containerConfigurator->import(PHPUnitSetList::PHPUNIT80_DMS);
-    $containerConfigurator->import(PHPUnitSetList::PHPUNIT_CODE_QUALITY);
-    $containerConfigurator->import(PHPUnitSetList::PHPUNIT_EXCEPTION);
-    $containerConfigurator->import(PHPUnitSetList::PHPUNIT_MOCK);
-    $containerConfigurator->import(PHPUnitSetList::PHPUNIT_SPECIFIC_METHOD);
-    $containerConfigurator->import(PHPUnitSetList::PHPUNIT_YIELD_DATA_PROVIDER);
-    $containerConfigurator->import(SetList::UNWRAP_COMPAT);
-    $containerConfigurator->import(SymfonySetList::SYMFONY_CODE_QUALITY);
-    $containerConfigurator->import(SetList::SAFE_07);
-    $containerConfigurator->import(SetList::TYPE_DECLARATION);
-
-    $parameters->set(Option::IMPORT_SHORT_CLASSES, false);
-
-    $parameters->set(
-        Option::SKIP,
-        [
-            __DIR__ . '/tests/bootstrap.php',
-            Rector\CodeQuality\Rector\Catch_\ThrowWithPreviousExceptionRector::class,
-            Rector\CodeQuality\Rector\Concat\JoinStringConcatRector::class,
-            Rector\CodeQuality\Rector\Identical\SimplifyBoolIdenticalTrueRector::class,
-            Rector\CodeQuality\Rector\Identical\SimplifyBoolIdenticalTrueRector::class,
-            Rector\CodingStyle\Rector\Class_\AddArrayDefaultToArrayPropertyRector::class,
-            Rector\CodingStyle\Rector\ClassConst\VarConstantCommentRector::class,
-            Rector\CodingStyle\Rector\ClassMethod\NewlineBeforeNewAssignSetRector::class,
-            Rector\CodingStyle\Rector\ClassMethod\RemoveDoubleUnderscoreInMethodNameRector::class,
-            Rector\CodingStyle\Rector\Encapsed\EncapsedStringsToSprintfRector::class,
-            Rector\CodingStyle\Rector\Switch_\BinarySwitchToIfElseRector::class,
-            Rector\CodingStyle\Rector\Use_\RemoveUnusedAliasRector::class,
-            Rector\Php56\Rector\FunctionLike\AddDefaultValueForUndefinedVariableRector::class, // Maybe good one day
-            Rector\PHPUnit\Rector\Class_\AddSeeTestAnnotationRector::class,
-            Rector\PHPUnit\Rector\ClassMethod\AddDoesNotPerformAssertionToNonAssertingTestRector::class,
-            Rector\TypeDeclaration\Rector\ClassMethod\AddArrayParamDocTypeRector::class,
-            Rector\TypeDeclaration\Rector\ClassMethod\AddArrayReturnDocTypeRector::class,
-        ]
-    );
+    $toSkip   = SerendipityHQ::buildToSkip(SerendipityHQ::SHQ_LIBRARY_SKIP);
+    $toSkip[] = \Rector\TypeDeclaration\Rector\Property\TypedPropertyFromAssignsRector::class;
+    $rectorConfig->skip($toSkip);
 };
