@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Serendipity HQ Value Objects Component.
  *
@@ -16,8 +18,6 @@ use SerendipityHQ\Component\ValueObjects\Address\Address;
 use SerendipityHQ\Component\ValueObjects\Address\AddressInterface;
 
 /**
- * {@inheritDoc}
- *
  * @ORM\Embeddable
  */
 class AddressEmbeddable extends Address
