@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Serendipity HQ Value Objects Component.
  *
@@ -25,14 +27,13 @@ final class AddressType extends AbstractType
 {
     /** @var string */
     private const LABEL = 'label';
+
     /** @var string */
     private const TRANSLATION_DOMAIN = 'translation_domain';
+
     /** @var string */
     private const SHQ_ADDRESS = 'shq_address';
 
-    /**
-     * {@inheritDoc}
-     */
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
