@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Serendipity HQ Value Objects Component.
  *
@@ -14,11 +16,10 @@ namespace SerendipityHQ\Component\ValueObjects\Money\Bridge\Doctrine;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
 use Doctrine\DBAL\Types\Type;
 use Money\Currency;
-use Money\Exception\ParserException;
-use Safe\Exceptions\StringsException;
-use function Safe\sprintf;
 use SerendipityHQ\Component\ValueObjects\Money\Money;
 use SerendipityHQ\Component\ValueObjects\Money\MoneyInterface;
+
+use function Safe\sprintf;
 
 /**
  * A custom datatype to persist a Money Value Object with Doctrine.
@@ -36,43 +37,38 @@ final class MoneyType extends Type
         return $platform->getVarcharTypeDeclarationSQL($column);
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function getDefaultLength(AbstractPlatform $platform): int
     {
         return 255;
     }
 
     /**
-     * {@inheritDoc}
-     *
-     * @throws \InvalidArgumentException
-     * @throws ParserException
      * @psalm-suppress MixedArgument
      *
-     * @return Money|string|null
+     * @return mixed|Money|string|null
      */
     public function convertToPHPValue($value, AbstractPlatform $platform)
     {
-        if (null === $value || '' === $value) {
+        if (null === $value || '' === $value || false === is_string($value)) {
             return $value;
         }
 
-        $objects = \explode('-', $value);
+        $objects  = \explode('-', $value);
+        $currency = $objects[1];
 
-        $currency = new Currency($objects[1]);
+        if (is_string($currency) && '' !== $currency) {
+            $currency = new Currency($currency);
+        }
+
+        if (false === $currency instanceof Currency) {
+            return null;
+        }
 
         return new Money([MoneyInterface::BASE_AMOUNT => (int) $objects[0], MoneyInterface::CURRENCY => $currency]);
     }
 
     /**
-     * {@inheritDoc}
-     *
      * @param Money|string|null $value
-     *
-     * @throws \InvalidArgumentException
-     * @throws StringsException
      *
      * @psalm-suppress DocblockTypeContradiction
      * @psalm-suppress MixedArgument
@@ -92,17 +88,11 @@ final class MoneyType extends Type
         return $value->getBaseAmount() . '-' . $value->getCurrency()->getCode();
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function requiresSQLCommentHint(AbstractPlatform $platform): bool
     {
         return ! parent::requiresSQLCommentHint($platform);
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function getName(): string
     {
         return self::NAME;
