@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Serendipity HQ Value Objects Component.
  *
@@ -37,9 +39,9 @@ final class VatNumberTest extends TestCase
         self::assertInstanceOf(VatNumberInterface::class, $resource);
         // Test the type of value object interface
         self::assertInstanceOf(ComplexValueObjectInterface::class, $resource);
-        self::assertEquals(self::TEST_DATA[VatNumberInterface::COUNTRY_CODE], $resource->getCountryCode());
-        self::assertEquals(self::TEST_DATA[VatNumberInterface::NUMBER], $resource->getNumber());
-        self::assertEquals(self::TEST_DATA[VatNumberInterface::VAT_NUMBER], $resource->getVatNumber());
+        self::assertSame(self::TEST_DATA[VatNumberInterface::COUNTRY_CODE], $resource->getCountryCode());
+        self::assertSame(self::TEST_DATA[VatNumberInterface::NUMBER], $resource->getNumber());
+        self::assertSame(self::TEST_DATA[VatNumberInterface::VAT_NUMBER], $resource->getVatNumber());
         self::assertIsString($resource->__toString());
         self::assertIsString($resource->toString());
     }
