@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Serendipity HQ Value Objects Component.
  *
@@ -36,8 +38,8 @@ final class VatRateTest extends TestCase
         self::assertInstanceOf(VatRateInterface::class, $resource);
         // Test the type of value object interface
         self::assertInstanceOf(SimpleValueObjectInterface::class, $resource);
-        self::assertEquals(self::TEST_DATA[VatRateInterface::COUNTRY_CODE], $resource->getCountryCode());
-        self::assertEquals(self::TEST_DATA[VatRateInterface::PERCENTAGE], $resource->getPercentage());
+        self::assertSame(self::TEST_DATA[VatRateInterface::COUNTRY_CODE], $resource->getCountryCode());
+        self::assertSame(self::TEST_DATA[VatRateInterface::PERCENTAGE], $resource->getPercentage());
         self::assertIsString($resource->__toString());
         self::assertIsString($resource->toString());
     }
